requirements.txt
streamlit
langchain
openai
python-dotenv
httpx
pycountry
langchain-community